fix: restore pull request checks
